Split the Traces V4 page-content test into two parallel suites
The single TracesV4PageContent.test.tsx had grown to 64 full-page-render
tests and ~685s wall-clock — the long pole of the js CI shard. Under CI's
--maxWorkers=2 the file ran serially on one worker, and late tests flaked on
their timeouts as the worker heap grew across dozens of prior renders (even a
3-row render tripped a 20s ceiling). Which test tripped was nondeterministic.

Extract the shared setup (msw server, render helper, mock state) into
tracesV4PageContentTestBed and split the describes across two sibling files so
CI's two workers run them concurrently. Mutable state is shared as stable
objects mutated in place (state, env) to survive the ESM import boundary; the
useExperimentQuery mock stays duplicated per file for babel hoisting.

No test logic, assertions, or names change. Each file now renders ~200s
locally instead of 685s, with headroom under the existing per-test timeouts.
